+++ djmoney/forms/fields.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from warnings import warn
-
 from django.core.exceptions import ValidationError
 from django.forms import ChoiceField, DecimalField, MultiValueField
 
@@ -17,27 +15,14 @@ __all__ = ('MoneyField',)
 
 class MoneyField(MultiValueField):
 
-    def __init__(self, currency_widget=None, currency_choices=CURRENCY_CHOICES,
-                 choices=CURRENCY_CHOICES, max_value=None, min_value=None,
-                 max_digits=None, decimal_places=None, default_amount=None,
-                 *args, **kwargs):
-
-        # choices does not make sense in this context, it would mean we had
-        # to replace two widgets with one widget dynamically... which is a
-        # mess. Instead, we let currency_choices be the same as choices and
-        # raise a warning.
-        if currency_choices != CURRENCY_CHOICES:
-            warn('currency_choices will be deprecated in favor of choices', PendingDeprecationWarning)
-            choices = currency_choices
-
-        # get the default currency if one was specified
-        default_currency = kwargs.pop('default_currency', None)
+    def __init__(self, currency_widget=None, currency_choices=CURRENCY_CHOICES, max_value=None, min_value=None,
+                 max_digits=None, decimal_places=None, default_amount=None, default_currency=None, *args, **kwargs):
 
         amount_field = DecimalField(
             *args, max_value=max_value, min_value=min_value, max_digits=max_digits, decimal_places=decimal_places,
             **kwargs
         )
-        currency_field = ChoiceField(choices=choices)
+        currency_field = ChoiceField(choices=currency_choices)
 
         # TODO: No idea what currency_widget is supposed to do since it doesn't
         # even receive currency choices as input. Somehow it's supposed to be