+++ djmoney/models/fields.py
@@ -276,7 +276,7 @@ class MoneyField(models.DecimalField):
     def formfield(self, **kwargs):
         defaults = {'form_class': forms.MoneyField}
         defaults.update(kwargs)
-        defaults['choices'] = self.currency_choices
+        defaults['currency_choices'] = self.currency_choices
         defaults['default_currency'] = self.default_currency
         if self.default is not None:
             defaults['default_amount'] = self.default.amount